+++ ld/writex86.c
@@ -43,7 +43,7 @@
 
 #define ABS_TEXT_MAX 64
 
-#define offsetof(struc, mem) ((int) &((struc *) 0)->mem)
+#define offsetof(struc, mem) ((size_t) &((struc *) 0)->mem)
 #define memsizeof(struc, mem) sizeof(((struc *) 0)->mem)
 
 PRIVATE bool_t bits32;		/* nonzero for 32-bit executable */
